Figurs

+++ vitalik/src/main/java/work/oop/figura/Triangle.java
@@ -4,13 +4,13 @@ package work.oop.figura;
  * Created by Виталий on 30.04.15.
  */
 public class Triangle extends Figura {
-    private int side;
+    private double side;
 
-    public int getSide() {
+    public double getSide() {
         return side;
     }
 
-    public void setSide(int side) {
+    public void setSide(double side) {
         this.side = side;
     }
